requirements.txt
falcon==3.1.1
falcon_auth==1.1.0
nzpy==1.15
python-dotenv==1.0.0
gunicorn==20.1.0
